+++ easy2homeassistant/easy_parser.py
@@ -1,281 +1,52 @@
 """A module to parse easy project xml files."""
 
-from dataclasses import dataclass, field
-from enum import Enum
-from typing import List, Optional
-
 import logging
 import xml.etree.ElementTree as ET
 
+from easy_types import Channel, Datapoint, Product, Project
+
 logger = logging.getLogger(__name__)
-
-
-# data structures
-class EntityKind(Enum):
-    """An enumeration to differentiate entities."""
-
-    UNDEFINED = 0
-    LIGHT = 1
-    COVER = 2
-    TEMPERATURE_SENSOR = 3
-    CLIMATE = 4
-
-    def __str__(self):
-        return f"{self.name} ({self.value})"
-
-
-@dataclass
-class Light:
-    """A data class to represent a light entity."""
-
-    name: str
-    address: int = 0
-    brightness_address: Optional[int] = None
-    state_address: int = 0
-    brightness_state_address: Optional[int] = None
-
-    def is_valid(self):
-        """Check if the entity is valid."""
-        return self.name != "" and self.address != 0 and self.state_address != 0
-
-    def get_kind(self):
-        """Return the entity kind."""
-        return EntityKind.LIGHT
-
-
-@dataclass
-class Cover:
-    """A data class to represent a cover entity."""
-
-    name: str
-    move_long_address: int = 0
-    stop_address: int = 0
-    position_address: int = 0
-    angle_address: int = 0
-    position_state_address: int = 0
-    angle_state_address: int = 0
-
-    def is_valid(self):
-        """Check if the entity is valid."""
-        return (
-            self.name != ""
-            and self.move_long_address != 0
-            and self.stop_address != 0
-            and self.position_address != 0
-            and self.angle_address != 0
-            and self.position_state_address != 0
-            and self.angle_state_address != 0
-        )
-
-    def get_kind(self):
-        """Return the entity kind."""
-        return EntityKind.COVER
-
-
-@dataclass
-class TemperatureSensor:
-    """A data class to represent a temperature sensor entity."""
-
-    name: str
-    state_address: int = 0
-    type: str = "temperature"
-    state_class: str = "measurement"
-
-    def is_valid(self):
-        """Check if the entity is valid."""
-        return self.name != "" and self.state_address != 0
-
-    def get_kind(self):
-        """Return the entity kind."""
-        return EntityKind.TEMPERATURE_SENSOR
-
-
-@dataclass
-class Climate:
-    """A data class to represent a climate entity."""
-
-    name: str
-    temperature_address: int = 0
-    target_temperature_state_address: int = 0
-    setpoint_shift_address: Optional[int] = None
-    setpoint_shift_state_address: Optional[int] = None
-    operation_mode_address: Optional[int] = None
-    operation_mode_state_address: Optional[int] = None
-    heat_cool_address: Optional[int] = None
-    heat_cool_state_address: Optional[int] = None
-    on_off_address: Optional[int] = None
-
-    def is_valid(self):
-        """Check if the entity is valid."""
-        return (
-            self.name != ""
-            # TODO and self.temperature_address != 0
-            and self.target_temperature_state_address != 0
-        )
-
-    def get_kind(self):
-        """Return the entity kind."""
-        return EntityKind.CLIMATE
-
-
-@dataclass
-class Entities:
-    """A data class to represent a collection of entities."""
-
-    light: List = field(default_factory=list)
-    cover: List = field(default_factory=list)
-    sensor: List = field(default_factory=list)
-    climate: List = field(default_factory=list)
-
-    def add_entity(self, entity):
-        """Add an entity to the corresponding list."""
-
-        if isinstance(entity, Light):
-            self.light.append(entity)
-        elif isinstance(entity, Cover):
-            self.cover.append(entity)
-        elif isinstance(entity, TemperatureSensor):
-            self.sensor.append(entity)
-        elif isinstance(entity, Climate):
-            self.climate.append(entity)
-        else:
-            logger.critical("Invalid entity '%s'", entity)
 
 
 # xml parsing
 class XMLParser:
     """A class to parse an easy project xml file."""
 
-    LIGHT_ADDRESS_MAP = {
-        "On/Off": "address",
-        "Dim value": "brightness_address",
-        "On/Off status": "state_address",
-        "Dim value status": "brightness_state_address",
-    }
-
-    COVER_ADDRESS_MAP = {
-        "Up/Down": "move_long_address",
-        "Step/Stop": "stop_address",
-        "Position control": "position_address",
-        "Slat angle control": "angle_address",
-        "Position control status": "position_state_address",
-        "Slat angle control status": "angle_state_address",
-    }
-
-    SENSOR_ADDRESS_MAP = {
-        "Indoor temperature": "state_address",
-    }
-
-    CLIMATE_ADDRESS_MAP = {
-        "Indoor temperature": "temperature_address",
-        "Room temperature": "target_temperature_state_address",
-        "Setpoint shift": "setpoint_shift_address",
-        "Setpoint shift status": "setpoint_shift_state_address",
-        "Mode": "operation_mode_address",
-        "Mode status": "operation_mode_state_address",
-        "Heat/Cool": "heat_cool_address",
-        "Heat/Cool status": "heat_cool_state_address",
-        "On/Off": "on_off_address",
-    }
-
-    ADDRESS_MAP = {
-        EntityKind.LIGHT: LIGHT_ADDRESS_MAP,
-        EntityKind.COVER: COVER_ADDRESS_MAP,
-        EntityKind.TEMPERATURE_SENSOR: SENSOR_ADDRESS_MAP,
-        EntityKind.CLIMATE: CLIMATE_ADDRESS_MAP,
-    }
-
     def __init__(self):
-        self.entities = Entities()
-        self.entity = None  # currently parsed entity
-        self.address_attribute_name = None  # currently parsed address
-        self.products = {}
-
-    def add_entity(self):
-        """Add the current entity to the entities list."""
-        if self.entity is not None:
-            self.entities.add_entity(self.entity)
-
-        self.entity = None
+        self.project = Project()
 
     def parse_group_addresses(self, group_addresses):
-        """Parse group addresses and set the lowest address to the currently parsed entity."""
-        lowest_address = float("inf")
+        """Parse group addresses and append them to the currently parsed datapoint."""
         for config in group_addresses.findall("config"):
             address = config.get("name")
+
             logger.debug("Parse group address '%s'", address)
             try:
-                lowest_address = min(lowest_address, int(address))
+                numeric_address = int(address)
+                self.project.channels[-1].datapoints[-1].groupAddresses.append(
+                    numeric_address
+                )
             except ValueError:
                 logger.warning("Skip invalid groupAddress '%s'", address)
 
-        if lowest_address != float("inf"):
-            setattr(self.entity, self.address_attribute_name, lowest_address)
-            logger.info(
-                "'%s': Set attribute '%s' to '%s'",
-                self.entity.name,
-                self.address_attribute_name,
-                lowest_address,
-            )
-        else:
-            logger.info(
-                "'%s': No group address for '%s'",
-                self.entity.name,
-                self.address_attribute_name,
-            )
-
     def parse_datapoints(self, datapoints):
-        """Parse datapoints and map them to the corresponding entity attribute."""
-        address_map = self.ADDRESS_MAP[self.entity.get_kind()]
+        """Parse datapoints and set them on the currently parsed channel."""
+
         for config in datapoints.findall("config"):
+            self.project.channels[-1].datapoints.append(Datapoint())
+
             for prop in config.findall("property"):
                 if prop.get("key") == "name":
-                    datapoint_name = prop.get("value")
-                    if datapoint_name in address_map:
-                        # set attribute name and search for group addresses
-                        self.address_attribute_name = address_map[datapoint_name]
-                        if hasattr(self.entity, self.address_attribute_name):
-                            self.parse_config(config)
-                        else:
-                            logger.error(
-                                "'%s' has no attribute '%s'!",
-                                self.entity.name,
-                                self.address_attribute_name,
-                            )
-                    else:
-                        logger.info(
-                            "'%s': Skip unmapped datapoint '%s'",
-                            self.entity.name,
-                            datapoint_name,
-                        )
+                    self.project.channels[-1].datapoints[-1].name = prop.get("value")
+            # parse group addresses
+            self.parse_config(config)
 
     def parse_context(self, context):
-        """Parse a context element and set the entity name."""
-        if self.entity.name != "":
-            logger.debug(
-                "The entity name is already set: '%s'. Skip parsing name by serial number!",
-                self.entity.name,
-            )
-            return
-
-        serial_number = ""
+        """Parse a context element and set the serial number on the currently parsed channel."""
         for prop in context.findall("property"):
             if prop.get("key") == "product.serialNumber":
-                serial_number = prop.get("value")
+                self.project.channels[-1].serialNumber = prop.get("value")
                 break
-
-        if serial_number in self.products:
-            self.entity.name = self.products[serial_number]
-            logger.info(
-                "Set entity name to '%s' by serial number '%s'.",
-                self.entity.name,
-                serial_number,
-            )
-        else:
-            logger.error(
-                "No product name found for serial number '%s'.",
-                serial_number,
-            )
 
     def parse_config(self, config):
         """Generic parser for a config element."""
@@ -311,42 +82,17 @@ class XMLParser:
             self.parse_config(config)
 
     def parse_channel(self, channel):
-        """Parse a channel element and create an entity."""
-        name = ""
-        kind = EntityKind.UNDEFINED
+        """Parse a channel element."""
+        self.project.channels.append(Channel())
 
         for prop in channel.findall("property"):
             if prop.get("key") == "Name":
-                name = prop.get("value")
+                self.project.channels[-1].Name = prop.get("value")
             elif prop.get("key") == "Icon":
-                icon = prop.get("value")
-                if icon == "icon-shutter":
-                    kind = EntityKind.COVER
-                elif icon in ("icon-light", "icon-dimmer"):
-                    kind = EntityKind.LIGHT
-                elif icon == "icon-indoor_temperature":
-                    kind = EntityKind.TEMPERATURE_SENSOR
-                elif icon == "icon-heat_regul":
-                    kind = EntityKind.CLIMATE
+                self.project.channels[-1].Icon = prop.get("value")
 
-        if kind != EntityKind.UNDEFINED:
-            if kind in (EntityKind.COVER, EntityKind.LIGHT) and name == "":
-                logger.debug("Skip unnamed channel of kind %s", kind)
-                return
-
-            # create entity and search for attributes
-            if kind is EntityKind.COVER:
-                self.entity = Cover(name)
-            elif kind is EntityKind.LIGHT:
-                self.entity = Light(name)
-            elif kind is EntityKind.TEMPERATURE_SENSOR:
-                self.entity = TemperatureSensor(name)
-            elif kind is EntityKind.CLIMATE:
-                self.entity = Climate(name)
-
-            logger.info("Found entity '%s' of kind %s", name, kind)
-            for config in channel.findall("config"):
-                self.parse_config(config)
+        for config in channel.findall("config"):
+            self.parse_config(config)
 
     def parse_channels_xml(self, channels_xml):
         """Parse the Channels.xml file and return the entities."""
@@ -357,29 +103,15 @@ class XMLParser:
 
         for channel in root.findall("config"):
             self.parse_channel(channel)
-            self.add_entity()
-
-        return self.entities
 
     def parse_product(self, product):
         """Parse a product element. Store the name for the serial number."""
-        name = ""
-        serial_number = ""
+        self.project.products.append(Product())
         for prop in product.findall("property"):
             if prop.get("key") == "SerialNumber":
-                serial_number = prop.get("value")
-                if serial_number == "":
-                    logger.warning(
-                        "Found product without SerialNumber: %s", product.get("name")
-                    )
+                self.project.products[-1].serialNumber = prop.get("value")
             elif prop.get("key") == "product.name":
-                name = prop.get("value")
-
-        if serial_number != "":
-            logger.info(
-                "Found product '%s' with serial number '%s'", name, serial_number
-            )
-            self.products[serial_number] = name
+                self.project.products[-1].name = prop.get("value")
 
     def parse_products_xml(self, products_xml):
         """Parse the Products.xml file."""
@@ -391,17 +123,6 @@ class XMLParser:
         for product in root.findall("config"):
             self.parse_product(product)
 
-    def remove_invalid_entities(self):
-        """Remove incomplete entities from the entities list."""
-        self.entities.light = [
-            light for light in self.entities.light if light.is_valid()
-        ]
-        self.entities.cover = [
-            cover for cover in self.entities.cover if cover.is_valid()
-        ]
-        self.entities.sensor = [
-            sensor for sensor in self.entities.sensor if sensor.is_valid()
-        ]
-        self.entities.climate = [
-            climate for climate in self.entities.climate if climate.is_valid()
-        ]
+    def get_project(self):
+        """Return the parsed project."""
+        return self.project
